requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
flask==2.3.3
flask-socketio==5.3.6
elasticsearch==7.17.9
python-socketio==5.8.0
python-engineio==4.7.1
requests==2.31.0
python-dotenv==1.0.0